Outsourced objective functions into own files.
