+++ tests/integration.c
@@ -8,32 +8,13 @@
 
 #include <criterion/criterion.h>
 
+#include "objectives.h"
 #include "../penguin/main.h"
 #include "../HGWOSCA/main.h"
 
 #ifndef M_PI
 #define M_PI (3.14159265358979323846)
 #endif
-
-/*******************************************************************************
-  SAMPLE OBJECTIVE FUNCTIONS
-******************************************************************************/
-
-double sum_of_squares(const double* const args, size_t dim);
-double sum(const double* const args, size_t dim);
-double rastigrin(const double* const args, size_t dim);
-double rosenbrock(const double* const args, size_t dim);
-double sphere(const double* const args, size_t dim);
-double egghol2d(const double* const args, size_t dim);
-double schaf2d(const double* const args, size_t dim);
-double griewank(const double* const args, size_t dim);
-
-/*******************************************************************************
-  UTILITIES
-******************************************************************************/
-
-void print_solution(size_t dim, const double* const solution);
-double* filled_array(size_t length, double val);
 
 
 /*******************************************************************************
@@ -75,25 +56,6 @@ Test(hgwosca_integration, sum) {
 }
 
 
-/* Test(hgwosca_integration, rastigrin) { */
-/*   size_t wolf_count = 40; */
-/*   size_t dim = 10; */
-/*   size_t max_iter = 100; */
-/*   double* mins = filled_array(dim, -100); */
-/*   double* maxs = filled_array(dim, 100); */
-/*   double* solution = hgwosca(rastigrin, wolf_count, dim, max_iter, mins, maxs); */
-/*   print_solution(dim, solution); */
-/*   for(size_t idx = 0; idx < dim; idx++) { */
-/*     cr_assert(fabs(solution[idx]) < 0.001, "each dimension should be reasonably close " */
-/*               "to 0"); */
-/*   } */
-/*   cr_assert(fabs(rastigrin(solution, dim)) < 0.1, "objective should be minimised at 0"); */
-/*   free(solution); */
-/*   free(mins); */
-/*   free(maxs); */
-/* } */
-
-
 Test(hgwosca_integration, rosenbrock) {
   size_t wolf_count = 40;
   size_t dim = 3;
@@ -112,21 +74,40 @@ Test(hgwosca_integration, rosenbrock) {
 
 
 Test(hgwosca_integration, sphere) {
-  size_t wolf_count = 30;
-  size_t dim = 10;
-  size_t max_iter = 800;
-  double* mins = filled_array(dim, -100);
-  double* maxs = filled_array(dim, 100);
-  double* solution = hgwosca(sphere, wolf_count, dim, max_iter, mins, maxs);
-  for(size_t idx = 0; idx < dim; idx++) {
-    cr_assert(fabs(solution[idx]) < 0.1, "each dimension should be reasonably close "
-              "to 0");
-  }
-  cr_assert(fabs(sphere(solution, dim)) < 0.5, "objective should be minimised at 0");
-  free(solution);
-  free(mins);
-  free(maxs);
+size_t wolf_count = 30;
+size_t dim = 10;
+size_t max_iter = 800;
+double* mins = filled_array(dim, -100);
+double* maxs = filled_array(dim, 100);
+double* solution = hgwosca(sphere, wolf_count, dim, max_iter, mins, maxs);
+for(size_t idx = 0; idx < dim; idx++) {
+cr_assert(fabs(solution[idx]) < 0.1, "each dimension should be reasonably close "
+"to 0");
 }
+cr_assert(fabs(sphere(solution, dim)) < 0.5, "objective should be minimised at 0");
+free(solution);
+free(mins);
+free(maxs);
+}
+
+
+/* Test(hgwosca_integration, rastigrin) { */
+/*   size_t wolf_count = 40; */
+/*   size_t dim = 10; */
+/*   size_t max_iter = 100; */
+/*   double* mins = filled_array(dim, -100); */
+/*   double* maxs = filled_array(dim, 100); */
+/*   double* solution = hgwosca(rastigrin, wolf_count, dim, max_iter, mins, maxs); */
+/*   print_solution(dim, solution); */
+/*   for(size_t idx = 0; idx < dim; idx++) { */
+/*     cr_assert(fabs(solution[idx]) < 0.001, "each dimension should be reasonably close " */
+/*               "to 0"); */
+/*   } */
+/*   cr_assert(fabs(rastigrin(solution, dim)) < 0.1, "objective should be minimised at 0"); */
+/*   free(solution); */
+/*   free(mins); */
+/*   free(maxs); */
+/* } */
 
 
 // Function is too noisy
@@ -184,97 +165,6 @@ Test(hgwosca_integration, sphere) {
 /*   free(mins); */
 /*   free(maxs); */
 /* } */
-
-/*******************************************************************************
-  IMPLEMENTATIONS
-******************************************************************************/
-
-void print_solution(size_t dim, const double* const solution) {
-  for(size_t idx = 0; idx < dim; idx++) {
-    printf("Item %ld:\t%f\n", idx, solution[idx]);
-  }
-}
-
-double* filled_array(size_t length, double val) {
-  double * res = (double*)malloc(length * sizeof(double));
-  for(size_t idx = 0; idx < length; idx++) {
-    res[idx] = val;
-  }
-  return res;
-}
-
-// Optimal solution is 0s everywhere.
-double sum_of_squares(const double* const args, size_t dim) {
-  double sum= 0.0;
-  for(size_t idx = 0; idx < dim; idx++) {
-    sum += pow(args[idx], 2.0);
-  }
-  return sum;
-}
-
-// Optimal solution is -100s everywhere.
-double sum(const double* const args, size_t dim) {
-  double sum = 0.0;
-  for(size_t idx = 0; idx < dim; idx++) {
-    sum += args[idx];
-  }
-  return sum;
-}
-
-// Optimal solution is 0s everywhere.
-double rastigrin(const double* const args, size_t dim) {
-  double rstgnND = 10 * dim;
-  for(size_t idx = 0; idx < dim; idx++) {
-    rstgnND += pow(args[dim], 2.0) - 10 * cos(2.0 * M_PI * args[dim]);
-  }
-  return rstgnND;
-}
-
-/** Multidimensional Rosenbrock Function
- global minima f(x1,.....,xN) = 0 at (x1,.....,xN) = (1,.....,1)*/
-double rosenbrock(const double* const args, size_t dim){
-  double rNd =0.;
-  for (size_t idx=0; idx<dim-1;idx++)
-      rNd += ( 100*(pow( args[idx+1] - pow(args[idx],2),2)) + pow(1-args[idx],2) );
-return rNd;
-}
-
-/* Multi Dimensional Sphere Function
- global minima at f(x1,.....,xN) = 0 at (x1,......,xN) = (0,......,0)*/
-double sphere(const double* const args, size_t dim){
-  double sph = 0;
-  for (size_t idx=0; idx<dim ; idx++)
-      sph +=  pow(args[idx],2);
-return sph;
-}
-
-/**2D Eggholder Function
-	global minima at f(x,y) = -959.6407 at (x,y) = (512,404.2319)*/
-double egghol2d(const double* const args, size_t dim){
-  //assert(dim == 2);
-  //assert( fabs(args[0])<=512 && fabs(args[1])<=512 );
-  double eggh = -1*(args[1]+47)*sin(sqrt(fabs(args[0]+(args[1]+47))))
-                  		- args[0]*sin(sqrt(fabs(args[0]-(args[1]+47))));
-return eggh;
-}
-
-/** 2D Schaffer Function */
-double schaf2d(const double* const args, size_t dim){
-  double sch = 0.5 + (pow(cos(pow(sin(fabs(args[0]*args[0]-args[1]*args[1])),2)),2) -0.5)/
-  										pow( (1+0.001*((args[0]*args[0]) + (args[1]*args[1]))), 2 );
-return sch;
-}
-
-
-// Griewank function
-double griewank(const double* const args, size_t dim) {
-  double sum = 0, prod = 1;
-  for(size_t idx = 0; idx < dim; idx++) {
-    sum += pow(args[idx], 2) / 4000;
-    prod *= cos(args[idx] / sqrt(idx));
-  }
-  return sum - prod + 1;
-}
 
 
 /*******************************************************************************
